How to run this:
This is not the "complete" project, as in you can't download it and run it directly on android studio.  It has the src java files, and the xml files. You will probably have to create a new android project and copy paste the *.java files and the *.xml files yourself.  

Also, some parts of the code, such as the package names and import statements of importing packages, and the custom database package 'db', will have to be edited.

*.png files need not be downloaded